SW-7406 - Fix http cache clear regression.

+++ engine/Shopware/Components/HttpCache/Store.php
@@ -103,6 +103,11 @@ class Store implements StoreInterface
                 continue;
             }
 
+            // skip .gitkeep
+            if ($file->getFileName() === '.gitkeep') {
+                continue;
+            }
+
             unlink($file->getPathname());
             $result = true;
         }
